+++ CHANGELOG.md
@@ -1,11 +1,14 @@
-## 1.0.0 (2026-07-21)
-
 # Changelog
 
 All notable changes to the **BufBear** extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- Added the `bufBear.formatDocument` command and `bufBear.formatting.enabled` setting.
 
 ## [0.1.0] - 2026-07-21